refactor: 대회 신청 커스텀 훅 분리
