+++ components/organism/CompetitionPage/ApplyPage/ApplyForm/ApplyForm.tsx
@@ -1,25 +1,12 @@
 'use client';
 
 import Button from '@/components/core/Button/Button';
-import React, { useEffect, useState, useTransition } from 'react';
+import React from 'react';
 import type { UserData } from '@/@types/user';
 import ApplyFormContent from './ApplyFormContent';
-import { useFormState } from 'react-dom';
-import { applyCompetition } from '@/app/competition/[id]/apply/applyCompetition';
-import { useRouter } from 'next/navigation';
 import { AnimatePresence } from 'framer-motion';
 import Dialog from '@/components/core/Dialog/Dialog';
-
-export type ApplyState =
-  | {
-      status: 'success';
-      message: string;
-    }
-  | {
-      status: 'error';
-      message: string;
-    }
-  | null;
+import useApplyForm from '@/lib/hook/useApplyForm';
 
 const ApplyForm = ({
   userData,
@@ -28,20 +15,8 @@ const ApplyForm = ({
   userData: UserData;
   competitionId: number;
 }) => {
-  const router = useRouter();
-  const [state, formAction] = useFormState<ApplyState, FormData>(applyCompetition, null);
-  const [, startTransaction] = useTransition();
-  const [isError, setIsError] = useState(false);
-
-  useEffect(() => {
-    if (!state) return;
-
-    if (state.status === 'error') setIsError(() => true);
-
-    if (state.status === 'success') {
-      router.push(`/competition/${competitionId}/success/`);
-    }
-  }, [state, router, competitionId]);
+  const { state, formAction, startTransaction, isError, setIsError } =
+    useApplyForm(competitionId);
 
   return (
     <>
